Section 19 at 31/3

+++ E-Commerce_BL/DTOs/ProductDTOs/AddProductDTO.cs
@@ -8,7 +8,7 @@ public class AddProductDTO
 {
     public string Name { get; set; }
     public string Description { get; set; }
-    public double Price { get; set; } = 0;
+    public decimal Price { get; set; } = 0;
     public string PictureURL { get; set; }
     public int ProductTypeId { get; set; }
     public int ProductBrandId { get; set; }
