+++ E-Commerce_BL/DTOs/ProductDTOs/ReadProductDTO.cs
@@ -7,11 +7,9 @@ public class ReadProductDTO
     public int Id { get; set; }
     public string Name { get; set; }
     public string Description { get; set; }
-    public double Price { get; set; } = 0;
+    public decimal Price { get; set; } = 0;
     public string PictureURL { get; set; }
-    //public int ProductTypeId { get; set; }
     public string productType { get; set; }
-    //public int ProductBrandId { get; set; }
     public string productBrand { get; set; }
     public bool IsDelete { get; set; }
 }
