+++ E-Commerce_BL/DTOs/ProductDTOs/UpdateProductDTO.cs
@@ -7,7 +7,7 @@ public class UpdateProductDTO
     public int Id { get; set; }
     public string Name { get; set; }
     public string Description { get; set; }
-    public double Price { get; set; } = 0;
+    public decimal Price { get; set; } = 0;
     public string PictureURL { get; set; }
     public int ProductTypeId { get; set; }
     public int ProductBrandId { get; set; }
